+++ src/react/src/components/Contribute/ProductionLocationDialog.jsx
@@ -26,7 +26,6 @@ import DialogTooltip from './DialogTooltip';
 import ProductionLocationDialogFields from './ProductionLocationDialogFields';
 import {
     mainRoute,
-    searchByNameAndAddressResultRoute,
     contributeProductionLocationRoute,
     MODERATION_STATUSES_ENUM,
     PRODUCTION_LOCATION_CLAIM_STATUSES_ENUM,
@@ -143,8 +142,8 @@ const ProductionLocationDialog = ({
     const statusLabel = startCase(toLower(moderationStatus));
 
     const handleGoToMainPage = () => history.push(mainRoute);
-    const handleGoToSearchByNameAndAddressResult = () =>
-        history.push(searchByNameAndAddressResultRoute);
+    const handleGoToSearchByNameAndAddress = () =>
+        history.push(`${contributeProductionLocationRoute}?tab=name-address`);
 
     const deleteIcon =
         moderationStatus === MODERATION_STATUSES_ENUM.PENDING ? (
@@ -285,7 +284,7 @@ const ProductionLocationDialog = ({
                             <Button
                                 variant="contained"
                                 color="secondary"
-                                onClick={handleGoToSearchByNameAndAddressResult}
+                                onClick={handleGoToSearchByNameAndAddress}
                                 className={classes.button}
                             >
                                 Submit another Location